Changed type of balance to float

+++ library/src/main/java/com/step/bank/Account.java
@@ -2,9 +2,9 @@ package com.step.bank;
 
 public class Account {
   private final String accNumber;
-  private int balance;
+  private float balance;
 
-  public Account(String accNumber, int balance) throws MinimumBalanceException {
+  public Account(String accNumber, float balance) throws MinimumBalanceException {
     this.accNumber = accNumber;
     if(balance < 1000){
       throw new MinimumBalanceException();
@@ -12,7 +12,7 @@ public class Account {
     this.balance = balance;
   }
 
-  public int getBalance() {
+  public float getBalance() {
     return balance;
   }
